+++ pom.xml
@@ -27,7 +27,7 @@
 	<dependency>
 	    <groupId>org.springframework</groupId>
 	    <artifactId>spring-jdbc</artifactId>
-	    <version>5.3.20</version>
+	    <version>5.3.20</version> 
 	</dependency>
 	
 	<!-- 커넥션풀 :  db open/close 을 관리해준다 -->
