+++ src/main/java/com/green/store/service/StoreService.java
@@ -3,24 +3,27 @@ package com.green.store.service;
 import java.util.List;
 
 import com.green.store.vo.HavingWineVo;
-import com.green.store.vo.StoreVo;
-import com.green.store.vo.WineListVo;
 import com.green.store.vo.RegVo;
+import com.green.store.vo.StoreVo;
 
 public interface StoreService {
 
-	List<RegVo> getSearchList(RegVo vo);
-
+	// (영태)
 	void insertSearch(RegVo vo);
 
-	List<HavingWineVo> getWineList(StoreVo vo);
+	// (영태)
+	List<RegVo> getSearchList(RegVo vo);
 
-	void updateWine(WineListVo vo);
-
-	void insertStore(StoreVo vo);
-
+	// (병규)
 	StoreVo storelogin(StoreVo vo);
 
+	// (병규)
+	void insertStore(StoreVo vo);
 
-	
+	// (임규)
+	List<HavingWineVo> getWineList(StoreVo vo);
+
+	// (임규)
+	void updateWineList(HavingWineVo havingVo);
+
 }
