+++ src/main/java/com/green/store/vo/WineListVo.java
@@ -7,15 +7,13 @@ public class WineListVo {
 	private int w_amount;
 	private int w_price;
 	private int s_no;
-	private int wl_idx;
 	
 	public WineListVo() {}
-	public WineListVo(int w_no, int w_amount, int w_price, int s_no, int wl_idx) {
+	public WineListVo(int w_no, int w_amount, int w_price, int s_no) {
 		this.w_no = w_no;
 		this.w_amount = w_amount;
 		this.w_price = w_price;
 		this.s_no = s_no;
-		this.wl_idx  =  wl_idx;
 		
 	}
 	public synchronized int getW_no() {
@@ -42,21 +40,10 @@ public class WineListVo {
 	public synchronized void setS_no(int s_no) {
 		this.s_no = s_no;
 	}
-	
-	public synchronized int getWl_idx() {
-		return wl_idx;
-	}
-	public synchronized void setWl_idx(int wl_idx) {
-		this.wl_idx = wl_idx;
-	}
-	
 	@Override
 	public String toString() {
-		return "WineListVo [w_no=" + w_no + ", w_amount=" + w_amount + ", w_price=" + w_price + ", s_no=" + s_no
-				+ ", wl_idx=" + wl_idx + "]";
+		return "StoreListVo [w_no=" + w_no + ", w_amount=" + w_amount + ", w_price=" + w_price + ", s_no=" + s_no + "]";
 	}
-	
-
