+++ src/main/webapp/WEB-INF/include/nav.jsp
@@ -4,23 +4,23 @@
 <!-- 위 nav (비로그인)-->
 <c:choose>
   <c:when test="${ loginVo eq null and sloginVo eq null }">
-	<ul class="nav justify-content-end">
-	  <li class="nav-item">
-	    <a class="nav-link active" aria-current="page" href="/UserLoginForm"><b>고객로그인</b></a>
-	  </li>
-	  <li class="nav-item">
-	    <a class="nav-link" href="/StoreLoginForm"><b>가게로그인</b></a>
-	  </li>
-	</ul>
+   <ul class="nav justify-content-end">
+     <li class="nav-item">
+       <a class="nav-link active" aria-current="page" href="/UserLoginForm"><b>고객로그인</b></a>
+     </li>
+     <li class="nav-item">
+       <a class="nav-link" href="/StoreLoginForm"><b>가게로그인</b></a>
+     </li>
+   </ul>
   </c:when>
   <c:when test="${ loginVo ne null and sloginVo eq null }">
- 	<ul class="nav justify-content-end">
-	  <li class="nav-item">
-	  	<br>
-	    <b> ${ loginVo.u_name } 님 환영합니다 </b>
-	  </li>
-	  <li class="nav-item">
-	    <img src="/img/logon.png" class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+    <ul class="nav justify-content-end">
+     <li class="nav-item">
+        <br>
+       <b> ${ loginVo.u_name } 님 환영합니다 </b>
+     </li>
+     <li class="nav-item">
+       <img src="/img/logon.png" class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
           <ul class="dropdown-menu">
             <li><a class="dropdown-item" href="#"><b>신청한 시음회</b></a></li>
             <li><a class="dropdown-item" href="/UserFavoriteStores"><b>관심 매장</b></a></li>
@@ -30,8 +30,8 @@
             <li><hr class="dropdown-divider"></li>
             <li><a class="dropdown-item" href="/UserLogOut"><b>로그 아웃</b></a></li>
           </ul>     
-	  </li>
-	</ul>
+     </li>
+   </ul>
   </c:when>
   <c:when test="${ lginVo eq null and sloginVo ne null }">
     <ul class="nav justify-content-end">
@@ -42,7 +42,7 @@
      <li class="nav-item">
        <img src="/img/logon.png" class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
           <ul class="dropdown-menu">
-          	<li><a class="dropdown-item" href="/StoreCheck"><b>매장정보</b></a></li>
+             <li><a class="dropdown-item" href="/StoreCheck"><b>매장정보</b></a></li>
             <li><a class="dropdown-item" href="/StoreWineRegisterForm"><b>매장와인 등록</b></a></li>
             <li><a class="dropdown-item" href="/StoreWineManage"><b>보유와인 확인/수정</b></a></li>
             <li><a class="dropdown-item" href="/Store/SalesHistory"><b>판매기록</b></a></li>
@@ -79,5 +79,4 @@
   </li>
 
 </ul>
-
