+++ src/main/webapp/WEB-INF/views/user/purchasehistory.jsp
@@ -36,7 +36,6 @@
       background-color: #f2f2f2;
       } 
 </style>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
 </head>
 <body>
  <%@include file="/WEB-INF/include/nav.jsp" %>
